Add support for several data functions. Refactor half to be explicit
Added support for

- logical binary operations
- diagonal
- identity
- iota
- ireduce
- range
- triangle
- transpose
- transpose_inplace

The OpenCL backend compiles but it is probably broken for f16. I will
be testing that in another PR.

+++ src/backend/cuda/kernel/lookup.hpp
@@ -34,7 +34,7 @@ __global__ void lookup1D(Param<in_t> out, CParam<in_t> in,
     int en = min(out.dims[vDim], idx + THRD_LOAD * THREADS);
 
     for (int oIdx = idx; oIdx < en; oIdx += THREADS) {
-        int iIdx     = trimIndex(idxPtr[oIdx], in.dims[vDim]);
+        int iIdx     = trimIndex(static_cast<int>(idxPtr[oIdx]), in.dims[vDim]);
         outPtr[oIdx] = inPtr[iIdx];
     }
 }
