+++ src/backend/opencl/magma/magma_blas_clblast.h
@@ -58,7 +58,9 @@ template <typename T> typename CLBlastType<T>::Type inline toCLBlastConstant(con
 // Specializations of the above function
 template <> float inline toCLBlastConstant(const float val) { return val; }
 template <> double inline toCLBlastConstant(const double val) { return val; }
-template <> cl_half inline toCLBlastConstant(const common::half val) { return val; }
+template <> cl_half inline toCLBlastConstant(const common::half val) {
+    return static_cast<float>(val);
+}
 template <> std::complex<float> inline toCLBlastConstant(cfloat val) { return {val.s[0], val.s[1]}; }
 template <> std::complex<double> inline toCLBlastConstant(cdouble val) { return {val.s[0], val.s[1]}; }
